Keep the cast that truncates a fraction of the registry
Not every `int(...)` is a uint cast. This one turns a float into a count,
since the caller asks for a fraction of the validators, and a range will
not take a float. The sweep read it as a cast and dropped it.

+++ tests/core/pyspec/eth_consensus_specs/test/helpers/rewards.py
@@ -409,7 +409,7 @@ def run_test_partial(spec, state, fraction_filled):
             data=state.previous_epoch_attestations[:num_attestations]
         )
     else:
-        for index in range(len(state.validators) * fraction_filled):
+        for index in range(int(len(state.validators) * fraction_filled)):
             state.previous_epoch_participation[index] = spec.ParticipationFlags(0b0000_0000)
 
     yield from run_deltas(spec, state)
